Mv implementation of From into impl
