+++ src/backend/src/impls.rs
@@ -1,7 +1,11 @@
-use crate::types::{Candid, StoredPrincipal};
+use crate::{
+    types::{Candid, StoredPrincipal},
+    State,
+};
 use candid::{CandidType, Deserialize, Principal};
 use core::ops::Deref;
 use ic_stable_structures::storable::{Blob, Bound, Storable};
+use shared::types::user_profile::Stats;
 use std::borrow::Cow;
 
 impl<T> Storable for Candid<T>
@@ -27,6 +31,16 @@ where
 
     fn deref(&self) -> &T {
         &self.0
+    }
+}
+
+impl From<&State> for Stats {
+    fn from(state: &State) -> Self {
+        Stats {
+            user_profile_count: state.user_profile.len(),
+            user_token_count: state.user_token.len(),
+            custom_token_count: state.custom_token.len(),
+        }
     }
 }
